Implementation of modal for adding information
