+++ react/src/components/skill.js
@@ -1,15 +1,23 @@
 import React from 'react';
-import {isAuthenticated,getSkill,delSkill} from '../user/auth';
-import { Collapse } from 'antd';
-import { CaretRightOutlined ,DeleteOutlined,BookOutlined,CalendarOutlined,BuildOutlined} from '@ant-design/icons';
+import {isAuthenticated,getSkill,delSkill,addSkill} from '../user/auth';
+import { Collapse,message,Modal,Popover } from 'antd';
+import { CaretRightOutlined ,DeleteOutlined,BookOutlined,CalendarOutlined,BuildOutlined,PlusCircleOutlined} from '@ant-design/icons';
 const { Panel } = Collapse;
 class Skill extends React.Component{
     constructor(props){
         super(props);
         this.state={
             id: this.props.match.params.id,
-            skill: []
+            skill: [],
+            title: "",
+            rate: "",
+            redirectToReferer: false,
+            error: "",
+            isModalVisible:false
         }
+    }
+    openModal(){
+        this.setState({isModalVisible:true});
     }
     getSkillDetails=async (id)=>{
         const data=await getSkill(id);
@@ -33,11 +41,62 @@ class Skill extends React.Component{
             this.getSkillDetails(this.state.id); 
         }
     }
+    handleCancel(){this.setState({isModalVisible:false})}
+    handleOk(){this.setState({isModalVisible:false},()=>{this.handleSubmit()})}
     componentDidMount(){
         this.getSkillDetails(this.state.id);
     }
     handleDelete=(eid)=>{
         this.deleteSkill(eid);
+        message.warning('Field Deleted from your profile');
+    }
+    handleChange=name=>event=>{
+        this.setState({
+            [name]: event.target.value,
+            error: "" 
+        }); 
+    };
+
+    isValid=()=>{
+       
+    }
+
+    handleSubmit=()=>{
+        const {title,rate}=this.state;
+
+        const user={
+            data:[
+            
+                {
+                title: title,
+                rate: rate
+                }]
+            
+            
+            }
+        
+        
+
+        const edit=async (id)=>{
+            const data=await addSkill(id,user);
+            console.log(data);
+            if(data.error){
+                this.setState({error: data.error});
+            }
+            else{
+                this.setState({
+                    title: "",
+                    descripton: "",
+                    error:"",
+                    redirectToReferer: true
+                });
+           }
+           window.location.reload();
+        }
+        
+        edit(isAuthenticated().user._id);
+
+    
     }
     render(){
         return(
@@ -59,17 +118,41 @@ class Skill extends React.Component{
                                  </div>
                                 </Panel>
                             </Collapse>
+                            <Modal title="Add Skill" visible={this.state.isModalVisible} onOk={()=>this.handleOk()} onCancel={()=>this.handleCancel()} okText='Add'>
+                            {this.state.error && 
+                                 message.error(this.state.error.toUpperCase())
+                            }
+                                <div className="container">
+                                    <div className="row">
+                                    <div className="col-md-4"></div>
+                                    <div className="">
+                                    
+                                    <div className="form-group">
+                                        <label><i className="fa fa-envelope" aria-hidden="true"></i> <b>Skill</b></label>
+                                        <input type="text" className="form-control" onChange={this.handleChange("title")} value={this.state.title}></input>
+                                    </div>
+                                    <div className="form-group">
+                                        <label><i className="fa fa-key" aria-hidden="true"></i><b>Rate [out of 10]</b></label>
+                                        <input type="number" className="form-control" onChange={this.handleChange("rate")} value={this.state.rate}></input>
+                                    </div>
+                                    </div>
+                               </div>
+                               </div>
+                            </Modal>
                         </div>
                     ))
                 }
+                
                  {isAuthenticated().user &&
                 <div className="" style={{marginTop: "20px",textAlign: "center"}}>
-                <button onClick={()=>this.props.history.push(`/editSkill/${this.state.id}`)} className="btn btn-warning" style={{backgroundColor: "#ffcc00"}}><i className="fa fa-plus" aria-hidden="true"></i> Add</button>
+                    <Popover placement="topLeft" content="Add Skill">
+                        <PlusCircleOutlined onClick={()=>{this.openModal()}}/>
+                    </Popover>
                 </div>
                 }
             </div>
         )
     }
-}
+} 
 
 export default Skill;
