--- react/src/editComponents/edit.css
@@ -1,12 +1,0 @@
-form{
-    box-shadow: 0px 0px 10px 1px #ffcc00
-}
-h2{
-    text-align: center
-}
-.btn{
-    background-color: #ffcc00;
-}
-h3{
-    margin-top: 20px
-}
